Solution to Simple Calculator

+++ Simple Calculator@@6788dd63490637cfc44e1ba8/code.c
@@ -7,19 +7,19 @@ int main(){
     scanf("%f %f",&a,&b);
     scanf("%c",&c);
     if(c=='+'){
-        p=int(a+b);
+        p=a+b;
         
         printf("%d",p);
     }else if(c=='-'){
-        p=int(a-b);
+        p=a-b;
         
         printf("%d",p);
     }else if(c=='*'){
-        p=int(a*b);
+        p=a*b;
         
         printf("%d",p);
     }else if(c=='/'){
-        p=int(a/b);
+        p=a/b;
         
         printf("%d",p);
     }else{
